Fix warnings issued by msvc for libs, pragmas and some casts (#2415)
- fix warnings issued by msvc for libs, pragmas, template definitions and some casts;
- introduce and make use of `DT_COMPILER_CLANG`, `DT_COMPILER_GCC` and `DT_COMPILER_MSVC` everywhere, except for the libs code; 
- rename `DTDEBUG` to `DT_DEBUG`.

WIP for #2413

+++ src/core/expr/head_literal_int.cc
@@ -104,7 +104,11 @@ Workframe Head_Literal_Int::evaluate_r(
         newcol = Const_ColumnImpl::make_int_column(1, value, col.stype());
       }
       else if (ltype == LType::REAL) {
-        newcol = Const_ColumnImpl::make_float_column(1, value, col.stype());
+        newcol = Const_ColumnImpl::make_float_column(
+                   1, 
+                   static_cast<double>(value), 
+                   col.stype()
+                 );
       }
       else {
         newcol = Const_ColumnImpl::make_int_column(1, value);
